Update ticky_check.py
I changed the csv file to have a function and a recursion system. That work very well with undoing the directory. I still need to test it. I need to figure out automating testing.  The for loop will go through the dict and take the values to the Per_user_list formatting.